\name{BH-package}
\alias{BH-package}
\alias{BH}
\docType{package}
\title{\packageTitle{BH}}
\description{\packageDescription{BH}}
\details{The DESCRIPTION file: \packageDESCRIPTION{BH}}
\section{Package Content}{\packageIndices{BH}}
\author{\packageAuthor{BH}}
\section{Maintainer}{\packageMaintainer{BH}}
\references{http://www.boost.org/}
\keyword{ package }
