\name{BH-package}
\alias{BH-package}
\alias{BH}
\docType{package}
\title{
The Boost C++ Libraries
}
\description{
Boost provides free peer-reviewed portable C++ source
  libraries.  We place them in this package as a more efficient
  distribution system for CRAN.
}
\details{
\tabular{ll}{
Package: \tab BH\cr
Type: \tab Package\cr
Version: \tab 1.51.0-2\cr
Date: \tab 2013-07-29\cr
License: \tab BSL-1.0\cr
}
Boost provides free peer-reviewed portable C++ source
  libraries.  We place them in this package as a more efficient
  distribution system for CRAN.  The idea is that you can write
  a package that depends on some subset of the Boost Libraries,
  and you won't need to distribute a copy of this code with your
  package.
}
\author{
John W. Emerson and Michael J. Kane and Dirk Eddelbuettel

Maintainers: Jay and Mike <bigmemoryauthors@gmail.com>, Dirk Eddelbuettel <edd@debian.org>
}
\references{
http://www.boost.org/
}
\keyword{ package }
\seealso{
\code{\link[bigmemory:bigmemory-package]{bigmemory}}
}
\examples{
# None
}
