\name{BH-package}
\alias{BH-package}
\alias{BH}
\docType{package}
\title{
  The Boost C++ Library Headers
}
\description{
  Boost provides free peer-reviewed portable C++ source
  libraries with a large part implemented as template headers.
  We place a number of them in this package as a more efficient
  distribution system for CRAN.
}
\details{
\tabular{ll}{
Package: \tab BH\cr
Type: \tab Package\cr
Version: \tab 1.55.0-3\cr
Date: \tab 2015-01-04\cr
License: \tab BSL-1.0\cr
}
  Boost provides free peer-reviewed portable C++ source
  libraries with a large part implemented as template headers.
  We place a number of these libraries in this package as a more efficient
  distribution system for CRAN.  The idea is that you can write
  a package that depends on some subset of the Boost Libraries,
  yet you do not need to distribute a copy of this code with your
  package.

  General discussions as well as questions and comments regarding the
  packages can be conducted at the mailing list for which subscriptions
  etc can be set up via the interface at R-Forge at
  \url{http://lists.r-forge.r-project.org/cgi-bin/mailman/listinfo/boostheaders-devel}.

  Bug reports can also be registered at the GitHub issue tracker at
  \url{https://github.com/eddelbuettel/bh/issues}.
}
\author{
John W. Emerson and Michael J. Kane and Dirk Eddelbuettel

Maintainer: Dirk Eddelbuettel <edd@debian.org>
}
\references{
http://www.boost.org/
}
\keyword{ package }
\examples{
# None
}
